feat(zkpvm): byte-wide AND table for the blake2b chips (C2)
The two blake2b chips proved every byte AND with two nibble lookups into
the 16x16 BitwiseLookupChip, carrying a hi-nibble witness column per
operand to split each byte. This adds a 2^16-row BitwiseAndByteChip
(a, b, a&b) with a free multiplicity column — byte-ness of a and b comes
free from table membership — and rewires both Blake2bChip and
Blake2bBoundaryChip (via the shared add_compression_core /
add_compression_interaction_core, generic over CompressionColumns) from
paired nibble lookups to single byte lookups. Each AND lookup halves; the
now-dead hi-nibble witness columns are deleted from both column enums.

Died from Column and BoundaryColumn (512 main limbs/row each):
  And{1,2,3,4}{A,B,Res}Hi (96), THi (16), AndT{Lo,Hi}Hi (16),
  HHi (64), VAfterHi (128), OutAnd1Hi (64), OutXor1Hi (64), OutAnd2Hi (64).
Survivors: And1..4, AndTLo/AndTHi, Output, OutAnd1, OutAnd2 (the AND
results, still fed to the arithmetic). The nibble table + its CpuChip /
ALU BitwiseChip consumers are untouched.

The byte table's multiplicity accumulates during blake2b trace-gen via
side_note.add_bitwise_and_byte (mirrors range256_counts); it is placed
last in BASE_COMPONENTS (a receiver, generated after all its consumers).
This changes the component set (chip_idx::COUNT 31 -> 32) and the
preprocessed shape, so the canonical profile and commitments move. Full
turnkey re-pin at the (32k, 8) cut: profile grows to [u32; 32] (chip 31 =
16, all others unchanged — floors still equal naturals), C_0
d76262877c32e2b6..., C_1 913f5cbf7cdd9d99...; VOUCHER_CHECK_* +
voucher-check catalog.toml re-pinned. prove.rs component_mask widened to
avoid a 1<<32 overflow at COUNT == 32.

Validated: blake2b_boundary_gate, memory_merkle_gate, chip_isolated
blake2b main+boundary (prove+verify), floors-cover, drift guard (C_0/C_1
reproduced), allowlist coverage (all 293 segments collapse to {C_0, C_1}),
catalog parity. Interaction-gen (release, synthetic): BLAKE2B_BOUNDARY
L16 617 -> 352 ms, BLAKE2B L13 177 -> 155 ms (on top of the dead-row
skip).

+++ zkpvm/src/prove.rs
@@ -577,8 +577,9 @@ pub fn prove_canonical(
     install_thread_pool();
     prepare_side_note_for_verification(side_note);
     let components = super::all_components();
-    // Full set present ⇒ constant mask = all 31 component bits set.
-    let component_mask = (1u32 << super::chip_idx::COUNT) - 1;
+    // Full set present ⇒ constant mask = all component bits set.  Built via
+    // a widening shift so COUNT == 32 (all u32 bits) doesn't overflow.
+    let component_mask = ((1u64 << super::chip_idx::COUNT) - 1) as u32;
     let (proof, _) = prove_impl_with_components_overridden(
         side_note,
         production_pcs_config_mobile(),
